Editor: moved content logic into a separate module
